Refactor: file 구조 변경
