+++ src/main/java/com/ceos20/spring_boot/chatting/repository/MessageRepository.java
@@ -1,6 +1,6 @@
-package com.ceos20.spring_boot.repository;
+package com.ceos20.spring_boot.chatting.repository;
 
-import com.ceos20.spring_boot.domain.ChattingRoom;
+import com.ceos20.spring_boot.chatting.domain.ChattingRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
