+++ src/main/java/com/ceos20/spring_boot/follow/repository/FollowRepository.java
@@ -1,6 +1,6 @@
-package com.ceos20.spring_boot.repository;
+package com.ceos20.spring_boot.follow.repository;
 
-import com.ceos20.spring_boot.domain.Follow;
+import com.ceos20.spring_boot.follow.domain.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
