+++ src/main/java/com/ceos20/spring_boot/post/domain/PostImage.java
@@ -1,10 +1,11 @@
-package com.ceos20.spring_boot.domain;
+package com.ceos20.spring_boot.post.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Column;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.FetchType;
 import lombok.AccessLevel;
@@ -13,8 +14,8 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostLike {
+@NoArgsConstructor (access = AccessLevel.PROTECTED)
+public class PostImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -23,7 +24,15 @@ public class PostLike {
     @JoinColumn(name = "post_id")
     private Post post;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+    @Column(nullable = false)
+    private String url;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String alt;
+
+    @Column(nullable = false)
+    private String type;
 }
