+++ src/main/java/com/ceos20/spring_boot/post/domain/PostLike.java
@@ -1,38 +1,30 @@
-package com.ceos20.spring_boot.domain;
+package com.ceos20.spring_boot.post.domain;
 
+import com.ceos20.spring_boot.user.domain.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
-import jakarta.persistence.Column;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.FetchType;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor (access = AccessLevel.PROTECTED)
-public class Follow {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class PostLike {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "follow_id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "following_user_id")
-    private User followingUser;
-
-    @Builder
-    public Follow(User user, User followingUser) {
-        this.user = user;
-        this.followingUser = followingUser;
-    }
 }
