+++ src/main/java/com/ceos20/spring_boot/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.ceos20.spring_boot.repository;
+package com.ceos20.spring_boot.post.repository;
 
-import com.ceos20.spring_boot.domain.Post;
+import com.ceos20.spring_boot.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
