+++ src/main/java/com/ceos20/spring_boot/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.ceos20.spring_boot.repository;
+package com.ceos20.spring_boot.user.repository;
 
-import com.ceos20.spring_boot.domain.User;
+import com.ceos20.spring_boot.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
